fixed: compile warnings (which picked up that CSysInfo::HasInternet was broken)
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@30700 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/StringUtils.cpp
@@ -270,7 +270,7 @@ CStdString StringUtils::SizeToString(int64_t size)
   CStdString strLabel;
   const char prefixes[] = {' ','k', 'M', 'G', 'T', 'P', 'E', 'Z', 'Y'};
   unsigned int i = 0;
-  double s = size;
+  double s = (double)size;
   if (s > 1000)
   {
     while (i < sizeof(prefixes)/sizeof(prefixes[0]) && s > 100.0)
